requirements.txt
attrs==25.1.0
aws-cdk-lib==2.180.0
aws-cdk.asset-awscli-v1==2.2.224
aws-cdk.asset-node-proxy-agent-v6==2.1.0
aws-cdk.cloud-assembly-schema==39.2.20
aws-cdk.lambda-layer-kubectl-v32==2.0.3
cattrs==24.1.2
cdk-ecr-deployment==3.0.99
constructs==10.4.2
importlib_resources==6.5.2
Jinja2==3.1.6
jsii==1.108.0
MarkupSafe==3.0.2
publication==0.0.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
six==1.17.0
typeguard==2.13.3
typing_extensions==4.12.2
